Refactor para Endpoint de create, update y delete review con OAuth2

+++ project/routers/reviews.py
@@ -8,6 +8,9 @@ from ..database import Movie
 from ..schemas import ReviewRequestModel
 from ..schemas import ReviewResponseModel
 from ..schemas import ReviewRequestPutModel
+# Aouth 
+from fastapi import Depends
+from ..common import get_current_user
 
 
 
@@ -17,20 +20,15 @@ router = APIRouter(prefix='/reviews')
 
 # Endpoint - Agregar reseña
 @router.post('', response_model=ReviewResponseModel)
-async def create_review(user_review: ReviewRequestModel):
+async def create_review(user_review: ReviewRequestModel, user: User = Depends(get_current_user)):
 
-    
-    if User.select().where(User.id == user_review.user_id).first() is None:
-        raise HTTPException(status_code=404, detail='El usurio no se ha encontrado')
 
     if Movie.select().where(Movie.id == user_review.movie_id).first() is None:
         raise HTTPException(status_code=404, detail='Pelicula no encontrada')
 
 
     user_review = UserReview.create(
-        # Las siguinetes variables son las creadas en el modelo 
-        # de validacion ReviewRequestModel
-        user_id=user_review.user_id,
+        user_id=user.id, # owner - propietario
         movie_id=user_review.movie_id,
         review=user_review.review,
         score=user_review.score
@@ -76,12 +74,15 @@ async def get_review(review_id: int):
 
 # Endpoint - Actualizar reseña 
 @router.put('/{review_id}', response_model=ReviewResponseModel)
-async def update_review(review_id: int, review_request: ReviewRequestPutModel):
+async def update_review(review_id: int, review_request: ReviewRequestPutModel, user: User = Depends(get_current_user)):
 
     user_review = UserReview.select().where(UserReview.id == review_id).first()
 
     if user_review is None:
         raise HTTPException(status_code=404, detail='Reseña no encontrada')
+    
+    if user_review.user_id != user.id:
+        raise HTTPException(status_code=401, detail='No eres el propietario')
     
     user_review.review = review_request.review
     user_review.score = review_request.score
@@ -96,12 +97,15 @@ async def update_review(review_id: int, review_request: ReviewRequestPutModel):
 
 # Endpoint Eliminar reseña 
 @router.delete('/{review_id}', response_model=ReviewResponseModel)
-async def delete_review(review_id: int):
+async def delete_review(review_id: int, user: User = Depends(get_current_user)):
     
     user_review = UserReview.select().where(UserReview.id == review_id).first()
 
     if user_review is None:
         raise HTTPException(status_code=404, detail='Reseña no encontrada')
+    
+    if user_review.user_id != user.id:
+        raise HTTPException(status_code=401, detail='No eres el propietario')
     
     user_review.delete_instance()
